Major refactoring of format package Implementation of harmony map supporting harmonize and flatten using rhino/javascript Retooled htmlresponse to inherit as delimited fixed some boxing/unboxing warnings other bugfixes and documentation updates

+++ yada-api/src/main/java/com/novartis/opensource/yada/util/YADAUtils.java
@@ -250,11 +250,11 @@ public class YADAUtils {
 				}
 				else if(param instanceof Integer)
 				{
-					p.setInt(i, (Integer)param);
+					p.setInt(i, ((Integer)param).intValue());
 				}
 				else if(param instanceof Float)
 				{
-					p.setFloat(i, (Float)param);
+					p.setFloat(i, ((Float)param).floatValue());
 				}
 			}
 			rs = p.executeQuery();
